+++ docs/CHANGELOG.md
@@ -26,6 +26,8 @@ _Latest build from master branch._
   ([#117](https://github.com/Tubeshade/Tubeshade/issues/117))
 - Correctly display rate for non-download tasks
   ([#234](https://github.com/Tubeshade/Tubeshade/pull/234))
+- Blocking task run ordering
+  ([#242](https://github.com/Tubeshade/Tubeshade/pull/242))
 
 ## [0.1.3] - 2025-12-28